requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
pandas>=1.5.0
openpyxl>=3.0.0
python-docx>=0.8.11
reportlab>=4.0.0
fake-useragent>=1.4.0
lxml>=4.9.0
selenium>=4.15.0